# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaKeKe API",
    "description": "TaKeKe API服务，提供认证和任务管理功能",
    "version": "1.0.0"
  },
  "paths": {
...
    "/tasks/": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "创建新任务",
        "description": "创建一个新的任务，支持设置标题、描述、状态、优先级、父任务等完整信息。",
        "operationId": "create_task_tasks__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "获取任务列表",
        "description": "获取任务列表 - 简化版本，只支持基本分页\n\n获取当前用户的任务列表，按创建时间倒序排列，支持分页和是否包含已删除任务。\n\nArgs:\n    page (int): 页码，从1开始\n    page_size (int): 每页大小，1-100\n    include_deleted (bool): 是否包含已删除的任务\n    user_id (UUID): 当前用户ID（从JWT token中获取）\n    session (Session): 数据库会话\n\nReturns:\n    TaskListResponseWrapper: 任务列表和分页信息响应\n\nRaises:\n    HTTPException: 查询参数错误或业务逻辑异常",
        "operationId": "get_task_list_tasks__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码，从1开始",
              "default": 1,
              "title": "Page"
            },
            "description": "页码，从1开始"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页大小，1-100",
              "default": 20,
              "title": "Page Size"
            },
            "description": "每页大小，1-100"
          },
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "是否包含已删除的任务",
              "default": false,
              "title": "Include Deleted"
            },
            "description": "是否包含已删除的任务"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "任务管理"
        ],
        "summary": "获取任务详情",
        "description": "获取任务详情\n\n根据任务ID获取任务的详细信息，包括所有字段和计算字段。\n\nArgs:\n    task_id (UUID): 任务ID\n    user_id (UUID): 当前用户ID（从JWT token中获取）\n    session (Session): 数据库会话\n\nReturns:\n    TaskGetResponse: 任务详情响应\n\nRaises:\n    HTTPException: 任务不存在或无权限访问",
        "operationId": "get_task_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "任务管理"
        ],
        "summary": "更新任务信息",
        "description": "更新现有任务的信息，包括标题、描述、状态、优先级等，支持部分更新。",
        "operationId": "update_task_tasks__task_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "任务管理"
        ],
        "summary": "删除任务",
        "description": "软删除指定任务，任务会被标记为已删除但不会物理删除，支持恢复操作。",
        "operationId": "delete_task_tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDeleteResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/complete": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "完成任务",
        "description": "完成任务并触发奖励分发。支持普通任务和Top3任务，奖励机制不同。",
        "operationId": "complete_task_tasks__task_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteTaskRequest",
                "examples": {
                  "NormalTaskCompletion": {
                    "summary": "普通任务完成",
                    "description": "完成普通任务，获得2积分基础奖励",
                    "value": {
                      "mood_feedback": {
                        "comment": "任务比较简单，顺利完成",
                        "difficulty": "easy"
                      }
                    }
                  },
                  "Top3TaskCompletion": {
                    "summary": "Top3任务完成",
                    "description": "完成Top3任务，有概率获得100积分或随机奖品",
                    "value": {
                      "mood_feedback": {
                        "comment": "挑战性任务，很有成就感",
                        "difficulty": "hard"
                      }
                    }
                  },
                  "TaskWithFeedback": {
                    "summary": "带详细反馈的任务完成",
                    "description": "完成任务并提供详细的情绪和难度反馈",
                    "value": {
                      "mood_feedback": {
                        "comment": "这个任务让我学到了很多新知识，虽然有些挑战但最终还是完成了",
                        "difficulty": "medium"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompleteResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/uncomplete": {
      "post": {
        "tags": [
          "任务管理"
        ],
        "summary": "取消任务完成",
        "description": "取消任务完成状态\n\n业务流程：\n1. 验证任务存在性和权限\n2. 检查任务是否处于完成状态\n3. 更新任务状态为pending\n4. 递归更新父任务完成度\n5. 记录操作日志\n6. 返回操作结果\n\n注意事项：\n- 取消完成不会回收已发放的积分或奖励，这是业务规则决定\n- 支持多层任务树的父任务完成度自动回退\n- 操作不可逆，请谨慎执行\n\nArgs:\n    task_id (UUID): 任务ID\n    request (UncompleteTaskRequest): 取消完成请求（空请求体）\n    user_id (UUID): 当前用户ID（从JWT token中获取）\n    session (Session): 数据库会话\n\nReturns:\n    TaskUncompleteResponseWrapper: 取消完成操作结果响应\n\nRaises:\n    HTTPException: 任务不存在、无权限访问或业务逻辑异常",
        "operationId": "uncomplete_task_tasks__task_id__uncomplete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UncompleteTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskUncompleteResponseWrapper"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompleteTaskRequest": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "CompleteTaskRequest",
        "description": "完成任务请求不需要请求体",
        "example": {}
      },
...
      "CreateTaskRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "description": "任务标题，1-100字符，必填",
            "example": "完成项目文档编写"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "任务描述，最多5000字符",
            "example": "编写项目的详细技术文档和用户手册"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "任务状态：pending/in_progress/completed",
            "default": "pending",
            "example": "pending"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "任务优先级：low/medium/high",
            "default": "medium",
            "example": "medium"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父任务ID，支持任务树结构"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "任务标签列表，最多10个标签",
            "default": [],
            "example": [
              "工作",
              "重要",
              "紧急"
            ]
          },
          "service_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Ids",
            "description": "关联服务ID列表，占位字段用于后续AI服务匹配",
            "default": []
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "任务截止日期（ISO 8601格式）"
          },
          "planned_start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Start Time",
            "description": "计划开始时间（ISO 8601格式）"
          },
          "planned_end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned End Time",
            "description": "计划结束时间（ISO 8601格式）"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateTaskRequest",
        "description": "创建任务请求Schema\n\n用于创建新任务的API请求。只包含创建任务时必要的字段，\n其他字段可以使用默认值或在创建后通过更新API设置。\n\n必填字段：\n- title: 任务标题（1-100字符）\n\n可选字段：\n- description: 任务描述\n- status: 任务状态（默认为pending）\n- priority: 任务优先级（默认为medium）\n- parent_id: 父任务ID\n- tags: 任务标签列表\n- due_date: 截止日期\n- planned_start_time: 计划开始时间\n- planned_end_time: 计划结束时间\n\n验证规则：\n- title长度必须在1-100字符之间\n- 如果设置parent_id，必须为有效的字符串ID\n- 如果设置时间范围，结束时间必须晚于开始时间\n- tags最多包含10个标签，每个标签最多20字符",
        "example": {
          "description": "编写项目的详细技术文档和用户手册",
          "due_date": "2024-12-31T23:59:59Z",
          "planned_end_time": "2024-12-30T18:00:00Z",
          "planned_start_time": "2024-12-20T09:00:00Z",
          "priority": "high",
          "status": "pending",
          "tags": [
            "文档",
            "项目"
          ],
          "title": "完成项目文档"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCreateResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "HTTP状态码"
          },
          "data": {
            "$ref": "#/components/schemas/TaskResponse",
            "description": "创建的任务信息"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          }
        },
        "type": "object",
        "required": [
          "code",
          "data",
          "message"
        ],
        "title": "TaskCreateResponse",
        "description": "任务创建响应\n\n继承自UnifiedResponse，data字段包含创建的任务信息。"
      },
...
      "TaskGetResponse": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "HTTP状态码"
          },
          "data": {
            "$ref": "#/components/schemas/TaskResponse",
            "description": "获取的任务信息"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          }
        },
        "type": "object",
        "required": [
          "code",
          "data",
          "message"
        ],
        "title": "TaskGetResponse",
        "description": "任务获取响应\n\n继承自UnifiedResponse，data字段包含获取的任务信息。"
      },
...
      "TaskListResponseWrapper": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "HTTP状态码"
          },
          "data": {
            "$ref": "#/components/schemas/TaskListResponse",
            "description": "任务列表和分页信息"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          }
        },
        "type": "object",
        "required": [
          "code",
          "data",
          "message"
        ],
        "title": "TaskListResponseWrapper",
        "description": "任务列表响应\n\n继承自UnifiedResponse，data字段包含任务列表和分页信息。"
      },
...
      "UncompleteTaskRequest": {
        "properties": {},
        "additionalProperties": false,
        "type": "object",
        "title": "UncompleteTaskRequest",
        "description": "取消任务完成请求不需要请求体",
        "example": {}
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}